HelpDesk_BE/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.14
h11==0.16.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.8.0
python-dotenv==1.1.1
python-jose==3.3.0
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
